Change LiveCounter to use Double instead of Int
When I copied the public API from JS in ce8c022 I didn't yet have enough
context to know how to translate the use of JS's `number` type. Now from
RTLc5 it's clear that counters are floating-point.

+++ Sources/AblyLiveObjects/DefaultRealtimeObjects.swift
@@ -38,7 +38,7 @@ internal final class DefaultRealtimeObjects: RealtimeObjects {
         notYetImplemented()
     }
 
-    internal func createCounter(count _: Int) async throws(ARTErrorInfo) -> any LiveCounter {
+    internal func createCounter(count _: Double) async throws(ARTErrorInfo) -> any LiveCounter {
         notYetImplemented()
     }
